feat(v3-sdk) export v3Swap logic
