+++ sdks/v3-sdk/src/utils/index.ts
@@ -11,6 +11,7 @@ export * from './nearestUsableTick'
 export * from './position'
 export * from './priceTickConversions'
 export * from './sqrtPriceMath'
+export * from './v3swap'
 export * from './swapMath'
 export * from './tickLibrary'
 export * from './tickList'
